Tweak database names in the tests.settings file for Django compatibility
This might look a little strange, but here's why I think it's the right
thing to do: If we want to have a persistent test database, we need to be
able to create it and apply migrations outside of a test session.
Unfortunately, Django doesn't seem to support either of those things via
its management tooling - it's only willing to create and migrate a test
database "inline" with respect to a test session.

Given that we need to create persistent test databases explicitly (via
createdb on the command line, say) this change makes everything downstream
of creation - incl. migrations and db fixtures - work as expected,
provided you specify the tests.settings module to Django (via "--settings"
or "--ds").