Tweaks to improve local test execution experience (#872)
* Update Pipfile (and lock) with missing dependencies

This patch synchronizes some tox and requirements.txt dependencies into
pipenv. Among other things, this makes it possible to run unit tests
without a tox environment.

* Remove load_cards Django command pytest fixture

This was taking ~20s per pytest session, whether the card data was already
loaded or not. Going forward, this should be run as part of the bootstrap
of a new test environment.

* Tweak database names in the tests.settings file for Django compatibility

This might look a little strange, but here's why I think it's the right
thing to do: If we want to have a persistent test database, we need to be
able to create it and apply migrations outside of a test session.
Unfortunately, Django doesn't seem to support either of those things via
its management tooling - it's only willing to create and migrate a test
database "inline" with respect to a test session.

Given that we need to create persistent test databases explicitly (via
createdb on the command line, say) this change makes everything downstream
of creation - incl. migrations and db fixtures - work as expected,
provided you specify the tests.settings module to Django (via "--settings"
or "--ds").

* Remove unused fake `hsredshift` database declaration in tests.settings

* Incorporate changes for new test workflow into tox.ini

Note that because we will be orchestrating the lifecycle of the
database(s) within tox, we'll need to remove `--create-db` from the Jenkins
build config.