[HSR-307] Manually refresh pending BillingAgreements for now
The PayPal API is currently omitting the billing_agreement_id on sales.
This means we're unable to update BillingAgreement's state if it isn't
in state="Active" right away upon executing the prepared agreement.

For now we're working around this by optimistically refreshing the
PayPal API if a user hits the billing page and has a pending
subscription. As subscriptions should only be in this state for a short
period of time this shouldn't lead to a large amount of API traffic.

This commit also reverts 16fe627af892daaf0cf6810a29dd011e51d056e2 with
the previous (non-functioning) attempt of using incoming webhook events
to refresh the billing agreement.

Refs paypal/PayPal-REST-API-issues#274.