Flake8 3.6 fixes
We previously misused the "# flake8: noqa" directive in an attempt to
disable checks for a single line. Instead, Flake8<3.6 disabled checks
for the full file. As Flake8 3.6 now requires the directive on a
separate line, a lot of files were now validated for the first time.

In addition this commit now disables W504 which is in conflict with
W503 (which we enforce).