Prevent one type of duplicate subscription
This could happen when a user was marked as past_due in Stripe, checked
out via PayPal through the Premium Modal, and the credit card was valid
the next time it retried.

The changes in this commit make the modal aware of past_due
subscriptions and prevent any new subscription, and also makes sure that
the billing callbacks do not create a new subscription when a Stripe one
exists.