Add grace period when rotating refresh tokens
This commit changes refresh token rotation to ensure that the
pre-rotation token can still be used within a grace period.

This may alleviate issues with mobile clients that attempt to request a
new access token, which causes HSReplay.net to rotate the refresh token,
but due to unstable network conditions the response never completes and
the old token is invalidated without the client being informed of the
new refresh token. The next time the client attempts to update it's
access token using the old request token it will be met with a client
error status code.

In order for clients to benefit they should retry fetching an access
token within the grace period which is currently set to 10 minutes.