Do not leak Uploads Tokens to third party OAuth clients
Previously, we only had two versions of the user serializer: the one
used by the GameReplaySerializer (with only username and battletag) and
the one used in the /api/v1/accounts/ endpoint.

As the /api/v1/accounst/ endpoint is critical for third-party apps to
discover the existing blizzard accounts (e.g. for consuming the
collection endpoint) we do not want to lock that endpoint down.

Rather, the endpoint now discriminates based on the authentication
mechanism which serializer to load: Authentication via Django session
includes all fields including some frontend-only fields, and OAuth
clients will receive different responses based on whether the access
token that was used for the request was granted the "fullaccess" scope.
In that case we use the same payload as previously. In all other cases,
there is now a new, limited serializer that only contains the username,
battletag, user id and their blizzard accounts.